[SES-4677] - Move avatar download out of worker and restrict concurrency (#1625)
* Limit concurrency on groups and communities

* Remove the use of worker for downloading avatar

* Adding mutex and concurrency limit on http requests
