+++ app/src/main/java/org/session/libsession/messaging/sending_receiving/pollers/OpenGroupPoller.kt
@@ -15,6 +15,8 @@ import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Semaphore
+import kotlinx.coroutines.sync.withPermit
 import org.session.libsession.database.StorageProtocol
 import org.session.libsession.messaging.jobs.BatchMessageReceiveJob
 import org.session.libsession.messaging.jobs.JobQueue
@@ -75,6 +77,7 @@ class OpenGroupPoller @AssistedInject constructor(
     private val communityDatabase: CommunityDatabase,
     @Assisted private val server: String,
     @Assisted private val scope: CoroutineScope,
+    @Assisted private val pollerSemaphore: Semaphore,
 ) {
     companion object {
         private const val POLL_INTERVAL_MILLS: Long = 4000L
@@ -97,7 +100,11 @@ class OpenGroupPoller @AssistedInject constructor(
 
             Log.d(TAG, "Polling open group messages for server: $server")
             emit(PollState.Polling)
-            val pollResult = runCatching { pollOnce() }
+            val pollResult = runCatching {
+                pollerSemaphore.withPermit {
+                    pollOnce()
+                }
+            }
             tokens.forEach { it.trySend(pollResult) }
             emit(PollState.Idle(pollResult))
 
@@ -424,6 +431,6 @@ class OpenGroupPoller @AssistedInject constructor(
 
     @AssistedFactory
     interface Factory {
-        fun create(server: String, scope: CoroutineScope): OpenGroupPoller
+        fun create(server: String, scope: CoroutineScope, pollerSemaphore: Semaphore): OpenGroupPoller
     }
 }
