+++ app/src/main/java/org/thoughtcrime/securesms/glide/RecipientAvatarDownloadManager.kt
@@ -1,8 +1,9 @@
 package org.thoughtcrime.securesms.glide
 
-import android.app.Application
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.debounce
@@ -12,13 +13,15 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.scan
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Semaphore
+import kotlinx.coroutines.sync.withPermit
 import network.loki.messenger.libsession_util.util.GroupInfo
 import org.session.libsession.utilities.TextSecurePreferences
 import org.session.libsession.utilities.recipients.RemoteFile
 import org.session.libsession.utilities.recipients.RemoteFile.Companion.toRemoteFile
 import org.session.libsignal.utilities.AccountId
 import org.session.libsignal.utilities.Log
-import org.thoughtcrime.securesms.attachments.AvatarDownloadWorker
+import org.thoughtcrime.securesms.attachments.AvatarDownloadManager
 import org.thoughtcrime.securesms.dependencies.ConfigFactory
 import org.thoughtcrime.securesms.dependencies.ManagerScope
 import javax.inject.Inject
@@ -28,11 +31,13 @@ import javax.inject.Singleton
 @OptIn(FlowPreview::class)
 @Singleton
 class RecipientAvatarDownloadManager @Inject constructor(
-    private val application: Application,
     private val prefs: TextSecurePreferences,
     private val configFactory: ConfigFactory,
     @ManagerScope scope: CoroutineScope,
+    private val avatarDownloadManager: AvatarDownloadManager,
 ) {
+    private val avatarBulkDownloadSemaphore = Semaphore(5)
+
     init {
         scope.launch {
             prefs.watchLocalNumber()
@@ -47,23 +52,38 @@ class RecipientAvatarDownloadManager @Inject constructor(
                         flowOf(emptySet())
                     }
                 }
-                .scan(State(emptySet())) { acc, newSet ->
+                .scan(State()) { acc, newSet ->
                     val toDownload = newSet - acc.downloadedAvatar
+                    val coroutineJobs = acc.downloadingJob.toMutableMap()
                     for (file in toDownload) {
                         Log.d(TAG, "Downloading $file")
-                        AvatarDownloadWorker.enqueue(application, file)
+                        coroutineJobs[file] = scope.launch {
+                            enqueueDownload(file)
+                        }
                     }
 
                     val toRemove = acc.downloadedAvatar - newSet
                     for (file in toRemove) {
                         Log.d(TAG, "Cancelling downloading of $file")
-                        AvatarDownloadWorker.cancel(application, file)
+                        coroutineJobs.remove(file)?.cancel()
                     }
 
-                    acc.copy(downloadedAvatar = newSet)
+                    acc.copy(downloadedAvatar = newSet, downloadingJob = coroutineJobs)
                 }
                 .collect()
             // Look at all the avatar URLs stored in the config and download them if necessary
+        }
+    }
+
+    private suspend fun enqueueDownload(file: RemoteFile) {
+        try {
+            avatarBulkDownloadSemaphore.withPermit {
+                avatarDownloadManager.download(file)
+            }
+        } catch (e: Exception) {
+            if (e !is CancellationException) {
+                Log.w(TAG, "Error downloading avatar $file", e)
+            }
         }
     }
 
@@ -113,7 +133,8 @@ class RecipientAvatarDownloadManager @Inject constructor(
     }
 
     private data class State(
-        val downloadedAvatar: Set<RemoteFile>
+        val downloadedAvatar: Set<RemoteFile> = emptySet(),
+        val downloadingJob: Map<RemoteFile, Job> = emptyMap()
     )
 
     companion object {
