+++ app/src/main/java/org/thoughtcrime/securesms/glide/RemoteFileLoader.kt
@@ -1,7 +1,5 @@
 package org.thoughtcrime.securesms.glide
 
-import android.content.Context
-import androidx.work.WorkInfo
 import com.bumptech.glide.Priority
 import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.Key
@@ -10,17 +8,13 @@ import com.bumptech.glide.load.data.DataFetcher
 import com.bumptech.glide.load.model.ModelLoader
 import com.bumptech.glide.load.model.ModelLoaderFactory
 import com.bumptech.glide.load.model.MultiModelLoaderFactory
-import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import org.session.libsession.utilities.recipients.RemoteFile
-import org.session.libsignal.exceptions.NonRetryableException
 import org.session.libsignal.utilities.Log
-import org.thoughtcrime.securesms.attachments.LocalEncryptedFileInputStream
-import org.thoughtcrime.securesms.attachments.AvatarDownloadWorker
+import org.thoughtcrime.securesms.attachments.AvatarDownloadManager
 import org.thoughtcrime.securesms.dependencies.ManagerScope
 import java.io.InputStream
 import java.security.MessageDigest
@@ -31,9 +25,8 @@ import javax.inject.Provider
  * A Glide [ModelLoader] for [RemoteFile]s
  */
 class RemoteFileLoader @Inject constructor(
-    @param:ApplicationContext private val context: Context,
     @param:ManagerScope private val scope: CoroutineScope,
-    private val localEncryptedFileInputStreamFactory: LocalEncryptedFileInputStream.Factory,
+    private val avatarDownloadManager: AvatarDownloadManager,
 ) : ModelLoader<RemoteFile, InputStream> {
     override fun buildLoadData(
         model: RemoteFile,
@@ -57,25 +50,7 @@ class RemoteFileLoader @Inject constructor(
         ) {
             job = scope.launch {
                 try {
-                    val downloadedFile = AvatarDownloadWorker.computeFileName(context, file)
-
-                    // Check if the file already exists in the local storage, otherwise enqueue a download and
-                    // wait for it to complete.
-                    if (!downloadedFile.exists()) {
-                        AvatarDownloadWorker.enqueue(context, file)
-                            .first { it?.state == WorkInfo.State.FAILED || it?.state == WorkInfo.State.SUCCEEDED }
-                    }
-
-                    val stream = localEncryptedFileInputStreamFactory.create(downloadedFile)
-
-                    if (stream.meta.hasPermanentDownloadError) {
-                        stream.close()
-                        throw NonRetryableException(
-                            "File download failed permanently for $file"
-                        )
-                    }
-
-                    callback.onDataReady(stream)
+                    callback.onDataReady(avatarDownloadManager.download(file))
 
                 } catch (e: CancellationException) {
                     Log.i(TAG, "Download cancelled for file: $file")
